+++ src/main/java/com/instantolap/charts/PositionAxis.java
@@ -2,8 +2,8 @@ package com.instantolap.charts;
 
 public interface PositionAxis extends Axis {
 
-  int getSamplePosition(Cube cube, int sample);
+  double getSamplePosition(Cube cube, int sample);
 
-  int getSampleWidth();
+  double getSampleWidth();
 
 }
